Grid controller
