+++ app/Http/Controllers/GridController.php
@@ -7,10 +7,23 @@
 
 namespace App\Http\Controllers;
 
+use Contracts\Repositories\PromoRepositoryContract;
 use Illuminate\Http\Request;
 
 class GridController extends Controller
 {
+    /**
+     * Construct the controller.
+     *
+     * @param PromoRepositoryContract $promo
+     * @param NewsRepositoryContract $news
+     * @param EventRepositoryContract $event
+     */
+    public function __construct(PromoRepositoryContract $promo)
+    {
+        $this->promo = $promo;
+    }
+
     /**
      * Display the grid view.
      *
