+++ contracts/Repositories/PromoRepositoryContract.php
@@ -10,4 +10,11 @@ interface PromoRepositoryContract
      * @return array
      */
     public function getHomepagePromos(int $page_id);
+
+    /**
+     * Get promotions for the grid.
+     *
+     * @return array
+     */
+    public function getGridPromos(int $page_id);
 }
